chore: improve VisuallyHidden exports
